+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 
 -   #2364 : Fix the use of the `--export-compiler-config` flag.
 -   #2372 : Fix passing arrays of size 0 to Fortran translations.
+-   #2306 : Fix Python containers as arguments to interface functions.
 -   \[INTERNALS\] Fix unsorted `__all__` variables.
 -   \[INTERNALS\] Allow CI scripts `check_pyccel_conventions.py`, `check_pylint_commands.py`, and `ci_tools/check_python_capitalisation.py` to be called easily locally.